Transform Component Introduced
- Position Component is now the transform component
- Transform component handles flipping, scaling, rotation (eventually),
and location
- Flipping functionality removed from graphicscomponent
- Render() no longer takes the position as a paramater, instead uses an
internal pointer to the transform component to find necessary attributes
- Transform component values can be initialised in the lua file

Elsewhere:
+ Fixed a bug where objects would move on window resize or moving.
_current speed was never 0 due to being a float, so there is now a kill
zone to stop negligible speeds amplifying
+ Removed the possibility of an endless loop. Instead of using recursion
to 'try again' in the Controller Component, I instead just copied the
contents of what the function should do. If there's no movement
component (which there shouldn't be), a single retry per input is
attempted along with a printf of the issue as notification.