Scripting is in :fire:
- A script component is defined with a path to a file elsewhere
- In this file, there can be an init, interact and update function.
These are ran in addition to other functions in the function lists
- Script components have a special constructor, giving them the ability
to run lua files due to their stored lua reference (_lua)
- Functions in LuaFunctions.h take an entity as a paramater (maybe more)
and call, return and do the appropriate things. Currently, printPhrase()
gets the npcComponent of an entity and prints the phrase stored within.
This works.
- Functions in LuaFunctions.h are loaded in loadFunctions(lua) at the
start of the program. This is binding c++ to lua, so hotloading doesn't
occur.
- Hotloading will need to be considered with the script components,
hence I store the file path name in the component as well.
- Functions in components will need to be put into the lfs namespace if
to be called from lua. These will be called after the entity is built,
so nothing to worry about
- Obviously, Entity.h now has a vector of void functions for
initialisation functions. This will mainly just contain the init script,
but may come in handy in the future
- Files are also restructured, where every entity has its own folder
unless its a common item

BUGS:
- Currently, only Init() works, but this will change!
- May need to look into how I pass Entity* _owner into a lua and whether
this is necessary every frame, how efficient it is etc