BIG CONVERSATIONAL UPDATE :fire:
No commits in a while, so here is something pretty big:

- Functions implemented to allow a scripter to manipulate the
conversation of any entity with an NPC component
- Functions called in Lua return true or false depending on whether they
were successful (try to keep this throughout the functions unless a
getter)
- The NPC component can have its speech functionality turned off (which
it is by default). This avoids the render call all together, no point
calling render when there's going to be no text on its way
- The text auto centers itself, but is all on one line for now. Of
course, putting a new line char works, but centers the whole thing.
Whether this is desired is not certain.
- No styling yet, but following what I've already done this will be
super easy to implement
- As with all lua helper functions, the entity to use is passed in, so
any NPC can be coversed with, even if the script is in another NPC. This
means I can make the enemy you damage in combat say "ow" or something
- Take a look at script.lua for basic conversation mechanics. May need
streamlining, but proves the point that through the power of scripting
we already have quite a flexible system

- The interaction prompt is now centered on the NPC, the idea being that
messages go above while the interact button stays visible
- The interact button can be toggled on and off, and is calibrated in
the constructor of the interaction component (off by default)
- You can still interact with something when there's no prompt - could
be used to end conversation implicitly or as a way of adding secret
interactive objects - spooky :ghost:
- Cleaned up a bit of code regarding positions and references to the
transform. Self explanitory.

- Cleaned up the code in the lua helper's "load functions", to split up
important parts. All functions must be loaded before the entities are
loaded, so the entity's init functions have the full pallette (while
there are already safeguards for this, its good semantically). Then the
execute now functions need all the entities loaded in order to play with
them.

That's all for tonight!