Implemented valid stats table (for HUD)
- This table should list every single stat in the game so that the
system knows which stats are going to be used by any combatant.
- So far, the game loads whether or not the stat should be displayed,
and if so, the style of appearance and any colours to show.
- Another system MUST be implemented to allow the modder to define HOW
the styles are drawn. Where are the bars drawn? How big? What order?
etc..
- Every colour that is added has a string attached (in an std::map), so
that when the HUD displays the stat it can apply the right colour to the
appropriate places